resolve-march-native: use gcc-unwrapped for detection
The wrapper that is used with the nixos gcc, does remove --march=native,
before handing over to gcc which means that the detection is not useable
out of the box and a user has to manually override the programs gcc with
its --gcc argument to an unwrapped gcc.
This patch does propagate the unwrapped gcc which prints out the actual
result without any customizations.