haskellPackages.extensions: use Cabal core pkg if possible
For GHC 9.6 and 9.8 we can use the Cabal library bundled with GHC since
it matches the requested version. This can prevent inconsistent
dependencies later in e.g. haskell-language-server.

For lower versions we may need to jailbreak and downgrade Cabal to which
upstream seems to be open:

> We need to pin a single major version of Cabal here because the main
> reason we use Cabal is for its list of extensions. Later versions have
> strictly more extensions, and we'll have missing patterns if we try to
> support more than one major version. If this causes problems in
> practice let's revisit this decision and come up with another
> approach.

Alternatively, hls-stan-plugin can be disabled.