haskell.packages.ghc98.haskell-language-server: allow build
Sorry, this is a bit of a draw the rest of the owl commit. Upgraded
where sensible/possible, all jailbreaks note the bounds issues which are
luckily relatively boring (base and friends).

Removed overrides are mostly stale overrides from configuration-head.nix
that don't work anymore and make little sense with the current package
set anyways.