python3.pkgs.pypaBuildHook: fix conflicts via propagated inputs 2
A follow up on #254547

As of now the hook still propagated `wheel` into the build environment which lead to a collision error as soon as any other version of wheel gets introduced as a dependency.

This makes the nixpkgs python infra more suitable for use with dynamic dependency versions like generated by 2nix tools.