qt5.qtbase: support substituting a cross-built qtbase dependency with a native-built one
For this we need to dynamically detect cross compilation in qtbase-setup-hook.sh
and in mkspecs, and stop unconditionally using binary lndir in propagatedNativeBuildInputs.

For now qtbase substitution will work only in one (most used) direction, and in many cases
will additionally require to manually add a buildHost version of qtbase.qmake to
nativeBuildInputs of the package that depends on qtbase.