firefox: set meta.mainProgram to launcherName in wrapper
After #235912 changed the binary name for some Firefox packages, the meta.mainProgram value was no longer accurate and commands such as nix run nixpkgs#firefox-devedition failed because /bin/firefox was executed instead of /bin/firefox-devedition.

This sets meta.mainProgram to launcherName, which contains the nameSuffix set in #235912, instead of inheriting the unwrapped package's mainProgram, which is always firefox with no suffix.