readline: Temporarily use older version of the patch
d6162d7857e76cf07fa5313451e7e8936f894594 did the "proper" work of fixing
the MinGW readline build, but won't work without
9cc3559a406209a55b4a78ae473214b378a8a029 which adds upstream-provided
patches but is only `staging-23.11`.

The older version of this patch however doesn't rely on p2-p7 so it will
work for now. `staging-23.11` will revert this because it does have
those patches.

The urgency is because this blocks Nix Windows development.