CODEOWNERS: add myself as code owner of the Nix package manager
Given the prior attempts to encourage Nix package bumps to be more stable, we failed, failed and failed again.

Today again, we shipped a broken Nix (2.19.2) without sufficient testing to unstable-small and it broke some users.

As announced in the past, I am taking ownership of the package manager bumps and modifications
as seemingly no one is interested to do it inside of Nixpkgs.