nix-cl: provide the original pname as an attribute
Overrides often inherit pname from the overriden package, and if they
get the already-prefixed version, defaulting the systems list to [pname]
produces wrong (and unintuitive) results. So we keep the prefix in the
store paths for clarity, and nix-env sees disambiguated names with -i/-u
(which should not be used anyway); but overrides won't be broken.