nixos/modemmanager: Add qmicli and mbimcli when using fccUnlockScripts
The ModemManager shipped fccUnlockScripts are using either qmicli or
mbimcli, to unlock wwan modems. These needs to be available for
ModemManager if you set the fccUnlockScripts link, so instead of having
every user than needs it pull it inn in different ways, we should just
provide the tools if we think they might be needed.

https://modemmanager.org/docs/modemmanager/fcc-unlock