packet-sd: backport outfile racyness/permission bugfix
The packet-sd exporter in version 0.0.3 creates a temporary file, moves
it to the target location and chmods it there. This creates a race
condition, in which Prometheus can be unable to read the file.

Additionally the outfile was created with too wide permissions (group-
and world-writable).

Both issues are resolved upstream, but not yet released.

(cherry picked from commit a9e9cb8cc686746e5c7d7741844f99c7afbcc263)