nixos/cage: document the default working directory
Most apps do not expect to be run from the system's root directory `/`. This PR
notifies users that the default working directory is `/`, and provides an
example for how to change that.

The motivation behind this PR is to avoid issues like
https://github.com/NixOS/nixpkgs/issues/229235#issuecomment-1891060190 and
https://github.com/NixOS/nixpkgs/issues/229235#issuecomment-1897512848. The
reason users need to be aware is to mitigate errors that are difficult to
diagnose, as they may (a) not be immediately attributable to the working
directory being the cause and (b) may only happen at certain events during
runtime.