auto-multiple-choice: use absolute paths instead of "portable" strategy
Currently, auto-multiple-choice from nixpkgs doesn't work on systems with
/lib directory. Running auto-multiple-choice there fails with:

    Unknown action gui at /nix/store/n3cyw5ngp0v52rny9gr6z35vslh5d4vx-auto-multiple-choice-1.6.0/bin/.auto-multiple-choice-wrapped line 94.

This happens on non-NixOS systems as well as on NixOS with e.g.
nix-ld [1] enabled.

The reason is documented in the Nix expression and the problem is
reported upstream at https://project.auto-multiple-choice.net/issues/872.
However, the problem can be worked around without requiring upstream
changes, as shown in this commit. We stop using the "portable
distribution", which tries to detect the base directory at run time, and
instead hardcode all paths as absolute at build time.

[1] https://github.com/Mic92/nix-ld