diffoscope: fix on aarch64-darwin
This was comparing stdenv.hostPlatform (an attr set) to a string. This
comparison always returned not-equal, which resulted in trying to
evaluate the broken gnumeric on macOS. This change fixes that to compare
the "system" string as was intended originally.

(cherry picked from commit 5d238d0d637c69c3a6fe77df87124c69192bdad3)