nixos/modules: fix mDNS name resolution
The underlying Avahi program was updated in 23.11, but it changed the
names of "mdns_minimal" to "mdns4_minimal" and "mdns6_minimal". This is
fixed in unstable, but for 23.11 it means there is a regression since
DNS name resolution for mDNS hosts (e.g., foo.local) won't work any
more.

This changeset fixes the issue by using the new configuration syntax.