antlr4.runtime.cpp: fix paths in generated cmakefile
The cmakefile for runtime config makes references to the binary outputs
but due to the package producing multiple outputs which separates the
library from includes, the path to the library directory is broken.

The library path has been corrected and the header path is also fixed
to avoid unnecessary hardcoding.