dockerTools: Add chown test
proot's --root-id "allows" chown only in the sense that it makes it
succeed vacuously, i.e. a no-op. This is undesired if the goal is to
actually create a layer with some files owned by different users.

Fortunately, fakeroot does allow persistence of emulated file owners,
and it is possible to combine fakeroot with proot, so replace proot
--root-id with fakeroot to do so.

This was fixed recently in d538fefb62a2dc0e40235606bb02615b47d02583,
so this commit just adds a test.