python3.pkgs.opentelemetry-instrumentation: 1.16.0 -> 0.43b0
This may seem like a downgrade, but actually it isn't:
the library tags releases as 0.Xb0 with the title "1.22.0 / 0.43b0".
Until 1.16 two tags were pushed, for 1.16 and 0.X. This isn't the case
anymore, so 0.43b0 is far newer than 1.16. This has practical
implications already since `opentelemetry-instrumentation-wsgi` doesn't
build at all:

    Checking runtime dependencies for opentelemetry_instrumentation_wsgi-0.37b0-py3-none-any.whl
      - opentelemetry-semantic-conventions==0.37b0 not satisfied by version 0.42b0

Originally, this package was published via 0.X, but this was changed
in 74303025237e4df97ba3597b089c1f1c596b9a71 which was a downgrade
already. This looks like a bug in the automatic python updater and
should be be skipped (or 1.x releases should be ignored).

ChangeLogs:
* https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.38b0
* https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.39b0
* https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.40b0
* https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.41b0
* https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.42b0
* https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.43b0

(cherry picked from commit c913a0bc0da401762630277afe845ff2cf8289a3)