libgit2: Enable tests
The tests hanging comment dates from 2018.
I did encounter a slow test when running them in a nix-shell -
the nonetwork::bad_urls one.
The name suggests that it might work fine, but it's still annoying
when building in nix-shell, and it seems unlikely that this
particular test will catch anything for us.