ghc: disable on powerpc64le
The first step of bootstrapping ghc on powerpc64le uses an ancient
ghc-8.6.5 binary which never worked very well and has always
segfaulted part of the time.  Unfortunately of late something in
newer Linux kernels, or maybe glibcs, or maybe something else is
causing it to segfault 100% of the time for me.

Sadly this means we can no longer build GHC on nixpkgs.  IMHO, long
term it just isn't sustainable for us to support compilers that
don't have either a bootstrap path from C/C++ or else a solid
self-cross-compilation-from-x86 bootstrap path.  I think we need to
wait for GHC to get the latter.

Also, recently people have been spraying ShellCheck far and wide
with wild abandon, which means that deliberately breaking GHC on
powerpc64le is the most straighforward way to get nixpkgs to skip
these ShellChecks (which of course need GHC).

So... this commit actually is necessary now in order to build the
NixOS installer.