swift: remove Concurrency backdeploy
This doesn't build with Swift 5.9 anymore, and sources were removed in
the Swift 5.10 branch:
https://github.com/apple/swift/commit/e54b18370d619eb766d953c2469109d03f9f7440

Because of this, I'm not sure if Swift code that uses Concurrency can be
built to target macOS <12 using our toolchain. That's very unfortunate,
but I don't see a way to remedy this. If I understand correctly, Xcode
relies on old binaries to accomplish this.