snapcast: fix build failure due to missing `#include <cstdint>`
`snapcast` fails to build with GCC 13 because of a missing include
of `cstdint` (that got transitively included in older versions).
This issue has already been fixed upstream some months ago, but
there have been no new releases after that, so this change pulls
in the upstream commit using `fetchpatch` for now.