Decrease the size of arm-none-eabi when using newlib-nano as libc.
Newlib-nano is used to decrease the produced binary size as much as
possible.

As such, it is usually built with -Os optimization flags
along with -fdata-sections and -ffunction-sections flags to allow
better discarding using `-Wl,--gc-sections`.

Libstdc++ in this case should also be built with the same flags and with
disabled exceptions in order to reduce it's binary size considerably.

The above match the behavior of the upstream GNU Arm Embedded toolchain.