Disable compiler runtime-test for cmake when cross-compiling to None platform.
CMake implicitly runs a compiler test when it detects a new project.
This normally compiles a minimal executable and runs it.

When cross-compiling to a None platform this will fail for all CMake packages
and the user is forced to override this manually through overlays.

This commit changes Cmake's behavior when cross compiling so that it
compiles a static library as a test instead.