nixos/samba: order after network-online.target
Reading the commit message on 50df42f0f314ca423ee299b8dc866fc3b189799b,
depending on network-online.target is definitely intentional and
potentially fixes real issues.

I *think* sticking it on the target should work fine, because the
services are brought in by the target, and if the target itself is
After=, that dependency relation should hold.