protobuf-c: unstable-2023-07-08 -> 1.5.0
Release notes:
https://github.com/protobuf-c/protobuf-c/releases/tag/v1.5.0

Split out `dev` and `lib` and added a workaround for the `unbound`
configure script, expecting both `include` and `lib` dirs in the
same base path.

Rerouted the unbound configure path for protobufc to the `dev`
subpackage.

Fixed anbox package moving protobufc to `nativeBuildInputs`, because
the `proto-c` tool is used at `build-time`.