grip-search: fix build with GCC 13
`grip-search` stopped building on GCC13 due to the standard
`cstdint` header not getting transitively included through
other standard headers anymore.

This change pulls in a patch from an open upstream PR [1]
that adds the necessary includes.

Since the current `substituteInPlace` patch that to the
built-in version string was overriding the `patches`
phase (preventing a simple declarative listing of patches
to be applied), I moved it to the `postPatch` hook instead.

[1]: https://github.com/sc0ty/grip/pull/6