nixos/apparmor: fix error on nixos-rebuild if no profiles are present
Restarting the apparmor service after a nixos-rebuild results in an
error on systems where no apparmor profiles are installed. This is
because the aa-remove-unknown program, which is called when the service
is reloaded, returns a non-zero return code if no profiles are present.

This changeset adds a guard to run aa-remove-unknown only if there are
profiles located in the /etc/apparmor.d directory, which resolves the
issue.

Resolves issue #273164.