arc_unpacker: apply AUR patch that fixes build on GCC 13
`arc_unpacker` fails to build with GCC 13 because GCC stopped
transitively including `cstdint` in a lot of scenarios.
While the project is inactive (the repository has been archived
on GitHub), the Arch Linux user repository already contains a
patch that fixes this issue.

Since ofborg indicated unit test failures on darwin-aarch64,
this change also marks the package as broken for such machines.
Note that these test failures are very unlikely to be caused
by the patch applied here, as it only adds a single
`#include <cstdint>` statement in a single file and that
specific include has been pulled in with older versions
of GCC, anyway.

Co-authored-by: midchildan <git@midchildan.org>
Co-authored-by: annalee <150648636+a-n-n-a-l-e-e@users.noreply.github.com>