aw-server-rust: fix failing build due to removed unstable feature
Rust compiler plugins have been entirely removed after having been
deprecated for years [1]. This caused the build of `aw-server-rust`
to break because its `aw-server` crate still contained the respective
crate attribute to enable the feature.

This attribute has already been removed upstream along with other
no longer used unstable feature opt-ins [2]. This change pulls in
the respective commit with `fetchpatch` and filters it down to
only the changes in `aw-server/src/lib.rs`, which fixes the build
for now. The patch should be removable with the next release
(0.12.3).

[1]: https://github.com/rust-lang/rust/pull/116412
[2]: https://github.com/ActivityWatch/aw-server-rust/commit/e1cd761d2f0a9309eb851b59732c2567a7ae2d3a.patch