doc: consistently use `nixpkgs-unstable` and "nixpkgs repository"
Rename channel `nixpkgs` to `nixpkgs-unstable`. Based on the
[repo branches](https://github.com/NixOS/nixpkgs/branches)
and [channel status](https://status.nixos.org) I don't believe there is
a `nixpkgs` channel.  This confused me as a Nixpkgs beginner.
Rename to `nixpkgs-unstable`, which does exist.

Refer to "nixpkgs repository" consitently.  Make the capitalization and
"code quoting" consistent when referring to the repository itself.