esphome: replace iputils dependency with inetutils.
The iputils package is only available on Linux. Adding it as a
dependency of esphome made the latter unavailable on other platforms.

The inetutils package also provides the ping command needed by esphome,
and is available on more platforms.

Also enable local networking on darwin in order for the tests to work in
the sandbox.