djv: fix build issue due to missing `cstdint` include
GCC 13 stopped transitively including a couple of headers (including
`cstdint`) in various scenarios, which breaks the build of the current
release version of `djv`. A fix for this issue has already been merged
in master as part of a PR from which Nixpkgs already pulls patches
from [1]. As this is the case, I just added the missing commit to
the list of patches in order to get the package building again.

[1]: https://github.com/darbyjohnston/DJV/pull/477/commits/be0dd90c256f30c0305ff7b180fd932a311e66e5