fbjni: fix broken build due to missing `cstdint` include
GCC 13 stopped transitively including `cstdint` in various scenarios,
breaking the build of `fbjni`. This issue has already been resolved
upstream but is currently only merged in `main` [1] (there hasn't been
a subsequent release yet).

This change pulls in the respective commit as a patch, fixing the
build until a new version of `fbjni` is relased.

[1]: https://github.com/facebookincubator/fbjni/commit/59461eff6c7881d58e958287481e1f1cd99e08d3