zotero: Add meta.mainProgram
The lib.getExe function's feature of assuming the program name from the
package is depracated, and it says to specify it here.

I was not sure whether I should do zotero itself, or .zotero-wrapped. I
put just zotero here, but I will change it if it should be the wrapped
version.

I didn't see any meta.mainProgam use the wrapped one, so I didn't use it
here.