libcamera-qcam: init
libcamera provides a more helpful debugging gui, which however needs qt,
so is undesirable in the default case (which is why it was disabled
initially).

Introduce a withQcam boolean flag, which will enable building it with
qcam, and expose it as a toplevel libcamera-qcam attribute.

This helped debugging the new ipu6 softisp camera stack.

The Qt 6 support patch is already circulating on the upstream mailing
list, however it doesn't apply cleanly. As discussed in
https://github.com/NixOS/nixpkgs/pull/284964, we opted to not ship
that patch, but rather use qt5 for now, as KDE seems to maintain a Qt5
patchset that we track.